Fully automated index and binding creation (#20)
# What

- When you create an index, the index table gets created automatically
- When you create a binding without a value for
`transformer_params["lexy_index_fields"]`, we’ll attempt to set the
value using `index_fields` from the corresponding index
- When a new index is created, we broadcast a signal to celery for pool
restart
  - The worker reloads index manager and is made aware of any new models


# Why

We can now build a full end-to-end workflow without having to restart
the app and without having to manually create db tables.


# Test plan

Create a new index with the following payload. The corresponding
database table `zzidx__my_test_index` will be created, and a signal will
be broadcast to restart the celery db worker.

```json
{
  "index_id": "my_test_index",
  "description": "My test index",
  "index_table_schema": {},
  "index_fields": {
      "word_count": {"type": "int"},
      "longest_word": {"type": "string","optional": true}}
}
```

Next, create a new binding with the following payload. The binding will
be created, and new tasks will be generated for all of the documents in
the `code` collection.

```json
{
  "collection_id": "code",
  "transformer_id": "text.counter.word_counter",
  "index_id": "my_test_index",
  "description": "Testing a new binding",
  "execution_params": {},
  "transformer_params": {},
  "filters": {}
}
```

Finally, retrieve the same binding created in the previous step. The
field `transformer_params` should include the key `lexy_index_fields`
and the value of the associated index fields from `my_test_index`.

```json
{
  "collection_id": "code",
  "transformer_id": "text.counter.word_counter",
  "index_id": "my_test_index",
  "description": "Testing a new binding",
  "execution_params": {},
  "transformer_params": {
    "lexy_index_fields": [
      "word_count",
      "longest_word"
    ]
  },
  "filters": {},
  "binding_id": 4,
  "status": "on",
  "created_at": "2023-10-31T21:07:25.249786+00:00",
  "updated_at": "2023-10-31T21:07:25.257910+00:00"
} 
```