Fixed bug in client that prevents creation of a binding without a filter (#44)
# What

Bugfix in the `Binding` client; passing `filter=None` when a filter is
not passed to `LexyClient.create_binding` method.

# Why

The client was passing `filter={}` when a filter was not set during
`create_binding` method (which is the default case). This caused the
server to attempt to validate `{}` as a filter which fails.

Not sure how this originally got past `examples/tutorial.ipynb`. Need to
start running tests through GH actions.

# Test plan

- [x] Run `pytest sdk-python`
- [ ] Run `examples/tests.ipynb` and verify that there are no errors
- [x] Run `examples/tutorial.ipynb` and verify that the tutorial works
as expected
- [ ] Run `examples/images.ipynb` and verify that the tutorial works as
expected