Refactor s3_bucket and s3_key (#116)
For storage objects, the `Document.meta` keys `s3_bucket` and `s3_key`
have been refactored to `storage_bucket` and `storage_key`.

This change does **not** include a migration since the fields are still
in `Document.meta` (for this exact type of flexibility). Therefore any
existing storage documents will have to be edited to include the new
keys.